Begin construction of the displayController class
I added an eventListener to the board container
that listens for any click. Upon click, I want to
store the first class of the target into a
variable. This way I can search my board array for
this particular class and change the display of
the corresponding button to the appropriate mark.