Add the ability to click on the board to mark
The cells displayed onscreen within the board
container now respond to the users clicks.

Currently, both the background color of the
cell and the board display printed to the console
change in response to a user click. The console
display will be used less and less as the game
is converted into the DOM.